backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0

# OpenAI
openai==1.12.0

# Export
openpyxl==3.1.2

# Environment
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
